Add ScriptedSubscriber
This commit introduces a new Subscriber implementation meant for testing
purposes: the ScriptedSubscriber. This subscriber is set up with signal
expectations before subscribing. Expectations will be verified with a
verification step, resulting in an AssertionError if they are not met.

Incorporates @nebhale's feedback.
Transferred from original PR in reactor/reactor-core#201