virtual time API split
This commit separates the virtual clock related methods into a dedicated
specialization of ValueBuilder so that they are only accessible when
explicitly needed.

The DefaultScriptedSubscriberBuilder does implement that specialized IF
but it is exposed as a "classic" ValueBuilder when instantiated via the
create() factory method, and a virtual-clock enabled builder when
instantiated via the withVirtualTime() factory method.