fix #32 Let StepVerifier work with simple VirtualTimeScheduler instance
This commit allows a StepVerifier to work with a naive Supplier for the
VirtualTimeScheduler where the enable method isn't called by the user.

VirtualTimeScheduler#enable can now take a pre-existing scheduler. The
operation is idempotent when attempted twice with the same instance.

Internally the StepVerifier default implementation will capture the vts
in the verify method and enable it.