fix #32 Ensure VirtualTimeScheduler is enabled when relevant
VirtualTimeScheduler#enable can now take a pre-existing scheduler. The
operation is idempotent when attempted twice with the same instance.
Note it may return a different scheduler if the currently enabled
scheduler is suitable.

Internally if a sourceSupplier was provided, the StepVerifier default
implementation will capture the vts in the verify method and enable it.
This means that any method of creating the vts will now work.

In the other case, the VirtualTimeScheduler supplier is still used but
the returned vts is not enabled. User is responsible for enabling it
beforehand and for ensuring the supplier always returns the same vts.
That scheduler is then manipulated by time-affecting methods.