Fix .expectNoEvents(duration).thenCancel()
When one or several `SubscriptionEvent` follow any `TaskEvent`, they
are converted to a `TaskEvent`. This TaskEvent will be correctly played
in sequence with other TaskEvent. This fixes the case where awaiting
would be short-circuited by a `thenCancel`. The conversion happens in a
conflateScript method.

Tests: conflateScript is tested, verifyVirtualTimeOnErrorAsync has been
fixed (it had a thenAwait wrongfully placed before thenRequest) and
verifyVirtualTimeOnNextIntervalReal has been made quicker.