Add thenConsumeWhile feature to StepVerifier
This commit adds a thenConsumeWhile method to StepVerifier that will
consume onNext signals until the given predicate stops matching (or a
termination signal is received).
At which point it resumes matching the signal with the subsequent
expectation.

The consumption of onNext is logged when using log().

First variant just discards the emitted items while the second one
allows to customize the consumption behavior.