fix #37 Verify that produced amount doesn't overflow request
This commit makes StepVerifier ensure that the produced amount doesn't
exceed the requested amount, if it is bounded.

An utility Processor has been created for the purpose of testing with a
publisher that misbehaves in that regard. It might be replaced in the
future by a more generic and publicly exposed API.