fix #101 expectNextSequence(empty)/Count(0) wrongly consumes an onNext
This commit fixes an issue where having an expectNextSequence(emptyIter)
or an expectNextCount(0) would consume the current onNext signal instead
of letting it through the rest of the scenario.

In these cases, the next expectation is pulled from the scenario and the
same signal is re-evaluated, thanks to the introduction of a reEvaluate
variable checked from `onNext`.