scheduler based on fork join pool (#113)
this is a Scheduler implementation that uses a ForkJoinPool and peer
Scheduler for scheduling time-based tasks.

it differs from the Parallel scheduler in that Workers are not pre-bound
to a specific thread in the pool. each Worker maintains its own queue of
tasks and when it has tasks to execute it will grab a thread from the
pool to use to execute it.

since the ForkJoinPool does not support time-based scheduling, another
Scheduler is used to enqueue tasks in the main FJ pool as necessary.