Tuple Decomposition Utilities (#115)
This change contributes a set of static methods that transform Consumers,
Functions, and Predicates of TupleN into Consumers, Functions, and Predicates
of N values, called with the contents of a Tuple.  This allows a function
written like this

.map(tuple -> {
  String name = tuple.getT1();
  String street = tuple.getT2();
  String city = tuple.getT3();
  String state = tuple.getT4();
  String zipcode = tuple.getT5();

  return String.format(...);
})

to instead be written as

.map(function((name, street, city, state, zipcode) -> String.format(...)))

and the equivalent for Consumers and Predicates.  These utility functions make
closures dealing with Tuples more expressive and less verbose.