[build] Bump wrapper to 4.2, update kotlin and shadowJar plugins
See reactor/reactor-core#915

This commit attempts at fixing the issue with sourceJar duplication,
now that kotlin plugin has a 1.1.60 version. After bumping the kotlin
plugin, the issue still manifested. Apparently the Gradle Wrapper was
still in version 2.8 and that had an impact.

Bumping gradlew to 4.2 and as a consequence shadowJar to 2.0.1 seems
to fix the issue, and the sourceJar can now be guarded against future
duplication issues by using a FAIL strategy.