Draft a tag and release note on non-snapshot publications (#245)
This change reads the version from `gradle.properties` and uses regexp
to interpret it, allowing the GitHub Actions workflow to determine if
we're dealing with a snapshot, a pre-release or a release.

In turn, it uses the later information to automatically prepare a draft
gh release for pre-releases and releases. Once the draft is manually
published, it will correctly tag the commit with `vVERSION` (light
tag as far as I understand, no comment but it should be fine).