[chores] Copyright: Correct start year for all files (#259)
This is a follow up on #258 which mistakenly harmonized all the start
dates in range to 2016. Some files have more ancient history than that,
having been transferred from older repositories.

This commit uses a mix of automation and manual process to refine the
start dates as much as possible.

As a result of automation, files that were missing the copyright notice
will have a range that reflects the git history, possibly with an end
date that is a past year (this has been fixed as a second step).

All ranges have then been reviewed and the start year amended when it
was necessary, and it was ensured that end year was also set to 2021.

This commit also manually adds copyright notices to kotlin files.

See #258.
See reactor/reactor#682.