Fix exponential Backoff overflow/npe issues by using ms for cap (#282)
This commit changes the exponential backoff default maxBackoff to be
Long.MAX_VALUE _milliseconds_ instead of _seconds_. The limit still
represents over 292 million years, plenty enough for a backoff :)

This prevents overflow and arithmetic exceptions, notably when computing
a jitter on top. This upper limit is also enforced in the constructor
when explicitly passing in a `maxBackoff`.

This commit also fixes 2 additional issues:
 - the computed BackoffDelay could have a `delay` component greater than
 the `max`, which would result in a negative jitter higher bound when
 computing the jitter bounds (now prevented in Backoff.exponential)
 - the "depends on previous value" variant would not use the default
 maxBackoffInterval if provided maxBackoff parameter was null, leading
 to a NullPointerException.

Supersedes and closes #269.